uint32_t to int32_t

+++ include/mxdrv.h
@@ -11,17 +11,17 @@
 #define MXDRV_STATUS_PAUSED         (0xFF00)
 #define MXDRV_STATUS_FADEOUT_END    (0xFF01)
 
-uint32_t mxdrv_load_mml(const uint8_t* mml_data, size_t mml_len, const uint8_t* data_title, int16_t use_pdx);
-uint32_t mxdrv_load_pcm(const uint8_t* pcm_data, size_t pcm_len, const uint8_t* pcm_name);
-uint32_t mxdrv_m_play();
-uint32_t mxdrv_m_end();
-uint32_t mxdrv_m_stop();
-uint32_t mxdrv_m_cont();
+int32_t mxdrv_load_mml(const uint8_t* mml_data, size_t mml_len, const uint8_t* data_title, int16_t use_pdx);
+int32_t mxdrv_load_pcm(const uint8_t* pcm_data, size_t pcm_len, const uint8_t* pcm_name);
+int32_t mxdrv_m_play();
+int32_t mxdrv_m_end();
+int32_t mxdrv_m_stop();
+int32_t mxdrv_m_cont();
 uint8_t* mxdrv_mml_name();
 uint8_t* mxdrv_pcm_name();
-uint32_t mxdrv_m_fadeout(uint16_t speed);
-uint16_t mxdrv_m_stat();
-uint16_t mxdrv_m_stat2();
-int16_t mxdrv_isavailable();
+int32_t mxdrv_m_fadeout(uint16_t speed);
+int32_t mxdrv_m_stat();
+int32_t mxdrv_m_stat2();
+int32_t mxdrv_isavailable();
 
 #endif
